refactor(appkit): extract shared type-generator error utilities
Move the connectivity-classification helpers (isConnectivityError,
getErrorDiagnostic, getErrorMessage and their internals) out of
query-registry.ts into a new errors.ts so both describe paths can share one
source of truth. Drop the duplicate errorMessageOf in statement-result.ts in
favour of the shared getErrorMessage.

Pure refactor — no behavior change.
